+++ packages/motion-dom/src/value/types/numbers/units.ts
@@ -1,6 +1,8 @@
+import { AnyResolvedKeyframe } from "../../../animation/types"
+
 /*#__NO_SIDE_EFFECTS__*/
 const createUnitType = (unit: string) => ({
-    test: (v: string | number) =>
+    test: (v: AnyResolvedKeyframe) =>
         typeof v === "string" && v.endsWith(unit) && v.split(" ").length === 1,
     parse: parseFloat,
     transform: (v: number | string) => `${v}${unit}`,